requirements.txt
Jinja2==3.0.3
PyYAML==6.0
requests==2.25.1
